microservices/OfferService/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
stomp.py==8.1.0
python-dotenv==1.0.0
aiohttp==3.9.1
pytest==7.4.3
httpx==0.25.1
motor==3.3.1
pymongo==4.6.0
dnspython==2.4.2
pytest-asyncio==0.21.1
pytest-cov==4.1.0
coverage==7.3.2 